requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyMuPDF==1.23.6
sentence-transformers>=2.5.0
chromadb==0.4.18
google-genai>=0.2.2
SQLAlchemy==2.0.23
streamlit==1.29.0
pytest==7.4.3
httpx==0.25.2
